Update ACS.asl
updated so its a bit more neat and that it can defiently start at certain parts. I also updated the pointers im using for stuff also fixed a "bug"(aka putting shit in the wrong place) so instead of the livesplit thinking your using steam instead of ubisoft i have no fix it so it knows.